requirements.txt
tensorflow>=2.10.0
opencv-python>=4.5.0
mediapipe>=0.8.0
pyspellchecker>=0.7.0
albumentations>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.0.0
numpy>=1.21.0 